# perf: save > 50% running time