--- a/__tests__/export.spec.ts
+++ b/__tests__/export.spec.ts
@@ -1,7 +1,9 @@
 import { name } from '../src';
+import { delay } from '../src/utils/delay';
 
 describe('jest-electron', () => {
-  test('export', () => {
+  test('async', async () => {
+    await delay();
     expect(name).toBe('jest-electron');
   });
 });