--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "jest-electron",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "description": "Easiest way to run jest unit test cases in electron.",
   "main": "index.js",
   "files": [