--- a/src/electron/interface.ts
+++ b/src/electron/interface.ts
@@ -1,3 +1,0 @@
-export type Args = {
-  readonly interactive?: boolean;
-}