--- a/src/utils/constant.ts
+++ b/src/utils/constant.ts
@@ -4,10 +4,5 @@ export enum EventsEnum {
   ProcRunTest = 'ProcRunTest',
   ProcRunTestResult = 'ProcRunTestResult',
   WebContentsReady = 'WebContentsReady',
-  WebContentsDestroyed = 'WebContentsDestroyed',
-  DevToolsOpened = 'DevToolsOpened',
   StartRunTest = 'StartRunTest',
-  EndRunTest = 'EndRunTest',
-  EndRunTests = 'EndRunTests',
-  ErrorRunTests = 'ErrorRunTests',
 }