--- a/src/utils/delay.ts
+++ b/src/utils/delay.ts
@@ -1,0 +1,11 @@
+/**
+ * 延迟 ms
+ * @param ms
+ */
+export const delay = (ms = 200): Promise<void> => {
+  return new Promise((resolve, reject) => {
+    setTimeout(() => {
+      resolve();
+    }, ms);
+  })
+};