使用 [lint-md](https://github.com/hustcc/lint-md) 的单测代码测试：

```
Test Suites: 40 passed, 40 total
Tests:       89 passed, 89 total
Snapshots:   0 total
```

结果如下：

- 1 main process

```
cache:
Time:        23.368s
no-cache
Time:        78.564s
```

- 2 main process

```
cache:
Time:        21.661s
no-cache:
Time:        77.522s
```

- multi-renderer

```
cache
Time:        8.458s
no-cache
Time:        42.851s
```

----

最后结论，multi-renderer 方式：

 - no-cache 情况下，运行时间降低到之前的 **54.5%**
 - cache 情况下，运行时间降低到之前的 **36.2%**